// src/urn.d.ts
declare const namespace: {
    binding: {
        redirect: string;
        post: string;
        simpleSign: string;
        artifact: string;
        soap: string;
    };
    names: {
        protocol: string;
        assertion: string;
        metadata: string;
        userLogout: string;
        adminLogout: string;
    };
    authnContextClassRef: {
        password: string;
        passwordProtectedTransport: string;
    };
    format: {
        emailAddress: string;
        persistent: string;
        transient: string;
        entity: string;
        unspecified: string;
        kerberos: string;
        windowsDomainQualifiedName: string;
        x509SubjectName: string;
    };
    statusCode: {
        success: string;
        requester: string;
        responder: string;
        versionMismatch: string;
        authFailed: string;
        invalidAttrNameOrValue: string;
        invalidNameIDPolicy: string;
        noAuthnContext: string;
        noAvailableIDP: string;
        noPassive: string;
        noSupportedIDP: string;
        partialLogout: string;
        proxyCountExceeded: string;
        requestDenied: string;
        requestUnsupported: string;
        requestVersionDeprecated: string;
        requestVersionTooHigh: string;
        requestVersionTooLow: string;
        resourceNotRecognized: string;
        tooManyResponses: string;
        unknownAttrProfile: string;
        unknownPrincipal: string;
        unsupportedBinding: string;
    };
};